Modifier-aware and deduplication key combination struct
Puts modifiers in a defined order at the start of the sequence of keys
and doesn't allow duplicates. Simply makes ordering the keys nicer and
more natural, similar to how it is usually displayed in hotkey
configuration software.
The key order doesn't match the addition order now if a modifier was
added, so the "undo" key is just a "pop" key. Sicne before this the keys
were just stored in order they were added regardless of value, "undo"
and "pop" were identical actions. Now the modifier keys are removed
first.

This could be improved by making each key a separate widget and have
each of them be removable on click, so the keys coudl be removed in
arbitrary order.